Sandy dev

+++ SweetDeal/SweetDeal/Models/College.swift
@@ -8,11 +8,11 @@ class College {
   var address: String
   var city: String
   var state: String
-  var lat: Double
-  var long: Double
+  var lat: Float
+  var long: Float
   var zip: Int
   
-  init(name: String, address: String, city: String, state: String, lat: Double, long: Double, zip: Int){
+  init(name: String, address: String, city: String, state: String, lat: Float, long: Float, zip: Int){
     self.name = name
     self.address = address
     self.city = city
